Description of this project
We have 4 person in our team. We won the first place of Datalink competition.
We considered how to predict COVID-19 confirmed cases with population mobility scenario.  Exploratory data analysis, feature analysis and model construction were done in this project. In exploratory data analysis part, we plotted the trends of confirmed covid cases and population mobility of several scenario. In feature analysis part, we analysed the correlation between population mobility of several scenario and confirmed covid cases. And we also did the variable importance by using random  forest to observe the sort of variable importance. Lastly, We only use KNeighborsRegressor model method to build the predictive mode because of limited time.